More "fun" with console colors; added three color schemes
Solarized, Monokai (via asciinema), and my own from an iMac called
Comonad.

To try these out, you have to have a truecolor-capable terminal.

Load the code and set truecolor mode:

  ./muforth -f lib/console-palette.mu4 +truecolor

and then print an example of one or more color schemes:

  solarized-original cr tt cc
  monokai cr tt cc
  comonad cr tt cc

tt and cc are just badly-named words that print some test text with
colors. I'm not sure why I called them that.

That's all. It's not that exciting.