[msp430] Improved text display on MSP430FR6989 Launchpad LCD
There is now a lookup table for ASCII characters 32 to 95. That's a
bunch of punctuation (not yet added), digits, and uppercase letters.
Later I'll add the segment bitmaps for the punctuation that it is
possible to represent on the LCD.

It's now easy to create Forth-style counted strings and display them -
either in the main LCD memory, the blinking memory, or both. I've
defined a few words to demonstrate this.

First you need to load everything. It's a mouthful:

  ./muforth -d fr6989 -f target/MSP430/load-kernel.mu4 -f target/MSP430/device/fr6989/lcd.mu4
  chat
  verify

If necessary, re-flash the kernel with

  flash-image
  verify

Then, start up the LCD:

  lcd-init

And try my demo words:

  howdy
  meow
  nums
  sneaker

Enjoy!