[msp430] Added a "load file" for MSP430FR6989 LCD code
To make it easier to experiment with the MSP430FR6989 Launchpad's LCD,
I've created a "load file" that loads everything you need.

Just do this:

  ./muforth -f target/MSP430/load-fr6989-lcd.mu4
  chat
  verify

and, if necessary,

  flash-image

All the code is burned into the Flash, so the entire RAM is free for
experimentation.

See earlier commits for ideas about things to try out.