[msp430] Make chat-v3 the default for MSP430FR6989
(This commit contains no changes to the tree, only edits to the previous
commit message, which contained both typos and errors.)

Because of the terrible USB-to-serial (USB-CDC) implementation in the
ezFET on the MSP430FR6989 Launchpad, I've written a version of the chat
code (v3) that has fewer round-trips between the host and the target.
This dramatically improves the latency when using this board.

However, currently chat-v3 isn't standard and you always have to specify
it. This commit changes that. The semantics are as follows:

* If you use  -d fr6989  on the command line (which "defines" fr6989)
and then load the chat firmware (using load-chat.mu4), by default you
will compile the newer chat v3 code.

* If you use  -d fr6989  on the command line for *any* code - including
loading the new chat code - by default you will get the *host-side
interface* for chat-v3.

* If you want to force the loading of chat-v3,  -d chat-v3
* If you want to force the loading of chat-v3-host,  -d chat-v3-host

* If you want to force chat-v2 or chat-v2-host, define either of those.

What this means for you:

If you have *never* loaded chat firmware onto your FR6989 Launchpad,
you'll need to get the code onto it *somehow*. I've used MSP430Flasher
(http://www.ti.com/tool/MSP430-FLASHER), which is clunky and requires
you to register at ti.com and agree to a ridiculous EULA to use, but
it's probably easier than cloning and building
https://github.com/dlbeer/mspdebug . Whichever poison you like, find
some.

Generate a .hex file with the compiled chat-v3 code:

  ./muforth -d fr6989 -f target/MSP430/load-chat.mu4 gen-ihex bye > chat-fr6989.hex

I haven't yet tried using mspdebug to do this, so you'll have to figure
that out.

For MSP430Flasher, after installing it, do something like this (assuming
$mspf is the directory containing the MSP430Flasher installation):

  LD_LIBRARY_PATH=$mspf $mspf/MSP430Flasher -w chat-fr6989.hex -v -z [VCC]

If you *have* installed chat firmware before (most likely chat-v2),
you'll have to do a two-step process to install chat-v3:

  ./muforth -d fr6989 -d chat-v2-host -d in-ram -f target/MSP430/load-chat.mu4
  chat
  go
  ^C

This will load the chat-v3 firmware into RAM and start executing it. The
^C will exit from muforth.

Now we load it again, but this time destined for the Flash:

  ./muforth -d fr6989 -f target/MSP430/load-chat.mu4
  chat
  flash-image
  verify
  ^C

verify  shows the addresses it's verifying. If there are no addresses
listed *after* the colons, then there were no mismatches.

You should reset the board after this so it will start executing the new
firmware from Flash.

At this point you should be able to load MSP430 code without having to
worry about chat-v2(-host) and chat-v3(-host). Everything should "just
work".