[msp430] Fixed the FR6989 CRC32 equates
I was surprised to see how messed up the documentation and equates both
are for the CRC32 peripheral.

The family user manual has garbled instructions on how to use the CRC32,
and a ton of registers, displayed in an unusual order. I'm pretty sure
there were some typos.

The datasheet (for the MSP430FR6989) shows many fewer registers, but
still shows signs of confusion. And there is at least one big typo:

CRC32RESRW1 is defined twice. The second time it should be listed as
CRC32RESRW0 instead.

I've fixed the typos (and confusion, I hope) and regenerated the
fr6989.mu4 equates file.