[msp430] Added chat v4 - for FR6989 and other FRAM devices
My reasons for creating yet another version of the chat protocol are
complicated, but the main reason was to try to get around the truly
awful "turnaround" latency of the ezFET USB-CDC serial interface on the
MSP430FR6989 Launchpad board.

I tried to make chat-v3 work with FRAM protection but ran into trouble,
and I was running out of space in the last 512 bytes of the FRAM too:
the vectors on this chip start at ff80 rather than ffc0 like on the
G2553.

In reimagining this version I tried to be as parsimonious as possible. I
like how it turned out.

It's much faster using an external serial connection - a PL2303 or
similar - but it's at least *usable* with the built-in ezFET serial
port.

I need to make some changes to existing code before I can tell you how
to switch to this. That is coming next.