[msp430] Don't force FR6989 LCD code to load into ram region
I thought I had removed this line.

In the load file for this code - mu/target/MSP430/load-fr6989-lcd.mu4 -
my intention was to load both the Forth kernel and LCD code into flash.

Having the LCD file say "ram" at the top prevents this.

What this change means to you is that you need to re-flash the kernel
and lcd code if you load it.