[msp430] Changed location of chat entry vector
It used to sit on the stack, pushed by the chat code when it started up.
But it took up a stack slot and was vulnerable to corruption.

This commit moves it to the start of the boot flash page. The 4 byte
Git commit that the chat code was built from has also been moved to the
start of flash, where verify can be smarter about ignoring changes to
it.

Because the chat entry vector is used by the Forth execution machinery,
any attempt to execute Forth code will fail until the chat code is
re-flashed. The kernel code (also in flash) does *NOT* need to change,
however. The trampoline code that calls into chat via the chat-entry
vector is in *RAM* and is copied to the target every time you "chat"
with the target.