[hc08] Fixes to BDM connection process
After the change to using chat-via, if the power-cycling process failed
to get the BDM target into debug mode, an error message would get
printed and the host would seem to "hang". It turns out it was waiting
for the target - which is *not* in debug mode - to return a valid status
- which is never going to happen.

I've changed the power-cycling process into a loop that the user can
exit at any time, but if successful connection *is* made, and the target
ends up in debug mode, then the target's registers are fetched and
displayed, and  du  and friends displays the target's memory rather than
that of the BDM "host" (the microcontroller driving the protocol).

With this change, using BDM works great. After not trying out this code
for several years, I was able to get a never-flashed S08QG into BDM,
and, after a few minutes of exploration, flash the newest S08 chat onto
it. BDM rules! It's much simpler and more reliable than TI's BSL. And
it's simple enough to implement in a small amount of code (though I
agonized for a long time writing the carefully-time bitbanging code).

Anyway, nice to go back in time and revisit this.