[core] Fixed lib/clashes.mu4
The word traversal code changed several times, and one of these broke
the clashes code.

I've fixed it, and I've also made it less confusing in its behaviour.
Now the order of the chains that are compared doesn't matter. In both
cases the chains are enumerated completely, all the way to the root,
through any linked chains. The order of the clashes that get printed
might differ, but the count and the words will be the same!