[s08] WIP: Spoken Word
Another work-in-progress checkin.

I breadboarded an S08QG8, an HC595, and 10 LEDs, and started to build a
prototype of my "Spoken Word" POV project. (See
https://www.nimblemachines.com/spoken-word/ for a few details.)

I decided midway through that it was time to finally get the S08 working
as a true Forth device, so I'm shelving this to focus on that.