[s08] Renamed target/HC08/ to target/S08/ (part 1)
The HC08 is dead, and very little of the code in target/HC08/ is truly
devoted to HC08/908 devices. It's almost exclusively HCS08/S08 devices.
I'd like the directory name to reflect this.

This commit is just the directory renaming. I have to go through and
edit all the local load files to reflect this change as well...

That's part 2.