[s08] Move to "standard" meta-compiler chains and token consumers
Take what has worked well for the ARM/V6-m, RISC-V, and MSP430 target
compilers and move it the S08.

There is still work to do - tweaks to meta-compiler to fix bugs and
clean it up - but remote execution works reasonably well now.

Existing kernel code shouldn't change with this commit, but is
guaranteed to in future commits. ;-)