[s08] Added a neat way to script target execution from the host
The (meta) word  script:  consumes a token from the input stream and
starts a special host colon compiler, with the following features:

  * target words are compiled as host literals, and later executed on
    the target
  * literal numbers become host literals (after target-specific
    conversion)
  * equates become literals
  * labels are *not* visible in this mode

The idea is that the same behaviour you get on the "chatting" command
line can be captured in host-side word and replayed.

Words defined with  script:  go into the .meta. chain.