[s08] Improvement to host-side BDM code
Be smarter about how the RAM-based flash programming routine is copied
to the target, and also smarter about *not* having du clobber that code;
dis *does* sync changes in ram to the target, but only under very
particular circumstances.

Also, when the BDM code first connects to the target, where it used to
look only for a factory trim byte, and give up if it was 0ff, now it
looks *first* for our custom trim byte (at 0ffad) - which trims the
oscillator to 9.216M - and uses that if found; then checks for the
factory trim, and only if both are missing does it complain to the user.