[s08] Even better BDM! v3 this time (sigh)
I wasn't sure I wanted to increment the version number. This is closer
to v1 than v2, and there really should only be *one* of these in the
world, so the best thing to do would be to make *the* versionless one
true BDM and get rid of v1 and v2...

Nevertheless, here we are.

If you want to use it, make sure to get the core routines loaded up on
the chip you want to use a "pod". Let's say that chip is a QG8:

  ./muforth -d QG8 -f target/S08/load-bdm.mu4
  chat
  flash-image
  verify
  ^C

Now find a chip to talk to (the "target"). Let's say this is a QG4.
Connect its BKGD pin (often pin 2) to the TPM(1)CH0 pin - PTA0 - on the
"pod". Find some code - maybe the chat code? - to load:

  ./muforth -d QG4 -f target/S08/load-chat.mu4
  bdm

Power cycle the target chip as instructed. You can now view memory,
flash and verify code, etc.

To watch the exciting chat and BDM serial protocols fly by, try saying

  spy on

first thing, or even put it on the command line at the end.

Enjoy!