[s08] Added "effective address" support in disassembler!
This is something that I've wanted for ever, and the underlying support
has been in the du/dis for a long time.

I just needed to "annotate" a few kinds of instructions and set the  ea
variable to the address referred to by the instruction.

How to use this: When disassembling, if you see a jump, branch, bsr, or
jsr, you can type "g" to "go" there, or "c" to call there. If you call,
you'll push a stack in the disassembler, and you'll get an extra ">" to
notify you of this. "r" pops the stack and returns you to the
instruction that you called from. It's very handy for following code,
esp when deeply nested.