[s08] JS16 USB-CDC code now enumerates and accepts buffers!
I've tested this on Linux but nowhere else.

This version of the USB-CDC code enumerates and creates a /dev/ttyACMx.
If you do this in another window:

  cat > /dev/ttyACMx

and then connect to the board like this:

  ./muforth -d JS16 -f target/S08/load-usb-cdc.mu4
  chat

you can spy on the text typed at it from the other window. Curiously, on
Linux anyway, a line of text typed to the terminal causes *two*
"packets" to be sent: first with all the text, then a two-byte packet
with CRLF in it.

out4-buffer is the endpoint buffer where the text shows up.

Sadly, at the moment I'm the only person who can see this, since JS16s
seem pretty rare.