[s08] Improvements to (JS16) USB chat code
For the one S08JS16 user - me. :-(

I've cleaned up the code, renamed a few things that were misleading,
moved code around, and fixed a bug: it used to be that zero-length
writes would fail (hang). The chip would still be chatting - it doesn't
run off into the weeds - but it meant that my code for copying the host
stack to the target for remote execution of Forth code (stack>) would
fail if the host stack was empty.

I could have fixed this by preventing the host from initiating
zero-length writes (zero-length reads work fine) but instead I thought
I'd try to solve the problem at the root. It took six more bytes of
code.

Anyhoo. the code is now as tight as I think I can get it. And I finally
understand completely how it works - I'm not "guessing" about it any
more. (Or so I hope.)