[core] Stack display is smarter and more crashproof
If a word underflowed the stack *and* called  error  or  error"  , the
error would be printed, but the stack depth remained negative. The (new)
stack printing code would try to print this strange stack and go into a
long loop, printing garbage and eventually seg faulting.

Stack display, and the error process, are both now smarter about this
situtation. I was unable to get muforth to crash.

I also added a stack depth indicator. When displaying the stack, a depth
is printed, followed by  =>  , followed by the stack items.