[arm] Got CMSIS-DAP working again, after mass erasing KL25Z!
Mass erase, requested from the MDM-AP, works as advertised: it erases
the flash, and sets the "secure bits" (byte 0000_040c) to fe (unsecure).

But having done this, a bit more is required to talk to chip out of
reset.

By trial and error (and tribulation) I was able to arrive at a sequence
that seems to work.