Simplified use of sed; we assume we can use -E
Instead of trying to figure out whether a platform has BSD or GNU sed
(or something else entirely, like the sed in busybox) we now assume that
the platform sed understand the POSIX -E flag, which changes the type of
regular expressions from "basic" to "extended".

If your sed can't do this, find one that can. GNU sed is one example
(but that is *not* an endorsement or even a recommendation!). And it may
have to be a newish GNU sed. I think I went through the rigamarole of
detecting which sed was available because BSD sed *supported* the POSIX
-E flag, while GNU sed did not (it used -r instead, but now supports
both).

I wonder whether busybox's or toybox's sed supports -E ?