riscv: [WIP] Beginnings of RV32C disassembler
The fiendish immediate encoding (in the assembler) and decoding (in the
disassembler) now seem to be able to correctly encode and decode each
bit for each type. So I'm pretty confident that they are correct.

Now I need to put the "shredder" pieces into the disassembler, and then
do some tests and point it at some real code (like the boot code in the
HiFive1's Flash ROM).