riscv: Make target/RISC-V/start-openocd.sh more flexible
Sometimes, esp when using prebuilt versions of openocd from SiFive! -
it's necessary to help tell openocd where to find copies of shared
libraries - esp libusb.

Now, in addition to the <openocd> variable that points to the openocd,
install there is a separate <libusb_dir> variable that points to a
directory containing a libusb.so.

Hopefully this makes things easier.