riscv: Improvements to FE310-G000 equates
I've decided to replicate, for each instantiated peripheral, the set of
register equates. In other words, for the three PWM units, there are not
three base addresses - PWM0, PWM1, and PWM2 - and one set of "offset"
equates that can be shared; instead, there is a full set of register
equates for *each* instantiated peripheral.

I think this reads better and will be easier to use.