riscv: [WIP] Experiments reading the SPI flash on HiFive1
This is a messy work-in-progress.

I was able to successfully read the JEDEC ID and to read bytes off the
chip.

For a while I was stuck thinking I had some kind of "off by 3" error: I
would read 4 bytes from 0000 but get four bytes from 0003. It turns out
that I was getting the right data, but the data I was comparing it to -
coming from SPI0 running in execute-in-place mode (XIP) was off. I had
somehow confused the state machine and it was reading the chip wrong!

After resetting the board and trying again, I realized I was getting
good data.

I need to figure out which erase command(s) to use, whether to flash by
byte or word, etc. More is coming.