riscv: Run your HiFive1 at 320 MHz!
I just wrote a bit of Forth code to experiment with the clocks on the
FE310/HiFive1.

The "led fade" demo uses the external oscillator to run the chip at 16
MHz. Since I've still got that code running on the board, this is what I
see when I fire up openocd.

Let's first see for sure that we are running at 16 MHz. (Run openocd in
another window, as usual.)

  ./muforth -f target/RISC-V/build.mu4 -f target/RISC-V/hifive1/clocks.mu4
  jtag
  speed

You should a value on the stack - in the ballpark of 16,000,000.

In addition to the external crystal oscillator (HFXOSC), the FE310 has
as an internal high-speed ring oscillator - HFROSC - that runs at a
nominal 72 MHz. At manufacture, a trim value is burned into the OTP
memory. Let's trim the HFROSC and see how fast we are running.

  >hfrosc
  speed  speed  speed

It's interesting to get several speed measurements. On my chip/board, I
got something like this:

jtag
      IX        SP        RP        IP
00000000  0000000b  00000000  00000000*  Ok (chatting)
>hfrosc
      IX        SP        RP        IP
00000000  80003f7c  80004000  80000890   Ok (chatting)
speed speed speed
      IX        SP        RP        IP
00000000  80003f78  80004000  80000890
      IX        SP        RP        IP
00000000  80003f74  80004000  80000890
      IX        SP        RP        IP
00000000  80003f70  80004000  80000890   Ok (chatting)
 3 =>          70005147          70005435          70011483

Right around 70 MHz.

Now let's fire up the PLL. We drive it from the HFXOSC. We have to
divide that 16 MHz clock to 8 MHz, since the PLL wants a reference clock
of 6 - 12 MHz. Then we multiply by 80 and divide by 2 to get 320 MHz!

  >pll
  speed speed speed

Again, on my board I get these numbers:

>pll
      IX        SP        RP        IP
00000000  80003f7c  80004000  80000890   Ok (chatting)
speed speed speed
      IX        SP        RP        IP
00000000  80003f78  80004000  80000890
      IX        SP        RP        IP
00000000  80003f74  80004000  80000890
      IX        SP        RP        IP
00000000  80003f70  80004000  80000890   Ok (chatting)
 3 =>         319998291         319998075         320003115