riscv: Be careful disassembling shift amounts
Previously I was using the same code that extracts rs2 (a 5-bit field)
to extract the (immediate) shift amount. For RV64, this is a 6-bit
field. (On RV32 it is as well, but the high bit is supposed to be zero.)

Now match it as a 6-bit field for both architectures.