riscv: [WIP] Trying to make a serial chat program in Forth
A tiny subset of the kernel. Trying to make it reasonably small.

It's currently a bit over 512 bytes - much bigger than it should be. I'd
like to see it be 256 bytes or less.

And it doesn't work. Command lookup table dispatch - which is klunky and
hard-to-read - seems to crash. The basic communication seems to work. I
can send a word over the wire and see it on the other end.