riscv: Made it easy to re-trim HFROSC
It turns out that on my HiFive1, the 72 MHz trim value isn't quite
right. The trim stored in OTP is 9. That yields a clock of about 70.7
MHz. A trim of 10 is much better, at 72.3 MHz.

You could test yours too. Load the clocks code (running openocd in
another tty session as usual):

  ./muforth -f target/RISC-V/build.mu4 -f target/RISC-V/hifive1/clocks.mu4
  jtag
  >hfrosc
  speed

This will push the speed of the HFROSC, trimmed with the factory trim
setting. You can see what it's currently trimmed to by doing:

  hfrosccfg @  16 >>  31 and

That will leave the current trim value on the stack.

To re-trim with a new value, just push the value and then do:

  retrim-hfrosc

then run

  speed

to see what happened.