riscv: Added a few multiply and divide words to kernel
Including some incomplete double-number code. It's not clear to me how
to synthesize a 64/32 division when the only primitive given is a 32/32
divide instruction. I tried something and then quickly gave up. I want
to revisit this, but at the moment it's on hold.

There are, however, perfectly working versions of

  *  /  um*  m*  /mod  u/mod

all checked in and tested. (um* and m* are easy, since RISC-V has mulh,
mulhu, and mulhsu instructions for calculating the high 32 bits of a
signed X signed, unsigned X unsigned, or signed X unsigned product,
resp.