riscv: Working chat code! (well, mostly)
I'm running into trouble because the code that is trying to execute
kernel words is building a stack frame in the same stack area that the
chat code uses for its stack.

So while the code works at first, after doing a few stack> and stack<,
it crashes.

Debugging this was *very* interesting. I used openocd to run the serial
chat command loop (code compiled to 8000_3000), and then left that
window open and started up muforth by loading the Forth kernel code,
connecting *via the HiFive1's debug serial port* (ttyUSB1), and talking
to the serial chat code. This code was compiled to 8000_0000.

Amazingly, this worked, and I was able to debug pretty seamlessly. Even
before the serial chat command loop was working, I could have it execute
a word that, say, sends a version number over the serial port. I could
then switch to the other muforth and "catch" this value - to make sure
that passing values back and forth was working. This was invaluable!!

I think I need to look at my MSP430 code and do something like the
"context pointer" that I have there.

But I definitely need to stop stepping on the stack toes of the chat
code!