riscv: Support for breaking 32-bit immeds into 20-bit and 12-bit parts
It's trickier than I thought. You can't simply split a 32-bit immediate
value into two fields, one 20 bits long and one 12 bits long. The
reason: the 12 bit field is treated as a *signed* value. If the low 12
bits represent a negative value, then the high 20 bits have to be
adjusted, so that their sum yields the original value.

It's not at all obvious, so I'm sure I'm not the first person who has
been surprised or bitten by this!