riscv: Trying to figure out why I can't program flash from chat
The chip always just hangs. Thinking that it might be related to somehow
jumping back into the flash space (for which the flash controller
returns values of 0000_0000 when it is in programmed i/o mode, which it
has to be in to program the flash), I decided that finally adding an
exception handler might help. By design (mostly likely), 0000_0000 is an
illegal instruction, and trying to execute this should cause an
exception.

If I execute a word with an ecall or 0000_0000 instruction, the chip
traps back into chat.

But if I jump to 2040_0000, the chip hangs. Not sure why this exception isn't
getting caught, or if it is, why we aren't ending up back in chat.