riscv: Working serial chat code; flashable from openocd or from chat!
It's not perfect, but it's good enough to poke around and get stuff
done, and be bothered restarting openocd all the time.

Here's how to get it onto your HiFive1:

Start openocd in another window. (Usual story: run
target/RISC-V/start-openocd.sh).

Load the chat code:

  ./muforth -d openocd -f target/RISC-V/load-chat.mu4
  jtag

The "-d openocd" tells muforth to expect to talk to the target using
openocd - hopefully for the last time!

Now for the big one:

  flash-image

It will print a few lines, saying what it's erasing and programming.
Afterwards you should verify the contents:

  verify

If verify prints nothing, then the image in the host's memory and on the
target's flash match perfectly. Reset your board by pressing the red
button, and quit muforth. We are about to connect with the HiFive1 using
the serial port that the FTDI chip makes available. On Linux it usually
shows up as /dev/ttyUSB1. On BSDs and OSX the names are different. But
it will be the *second* of two. The first one is the JTAG interface.

You need to make a symlink so that muforth can find the serial device.
In the muforth/mu directory, do something like this:

  ln -s /dev/ttyUSB1 serial-target

substituting your actual device for ttyUSB1, and reload muforth like this:

  ./muforth -f target/RISC-V/load-chat.mu4

This will load the same code as before.

Now, instead of typing "jtag" to connect, type

  chat

You'll see the chat version commit and a register dump. Now you should
be able to execute Forth code and poke around as before. But every time
you reboot/reset your board, it will be running chat code that is easy
to re-connect to. As long as you don't say "-d openocd", muforth will
expect to see a serial chat target.

As always, enjoy!