[core] Tweaks to markup to HTML code
Still need to figure out bulleted lists, tables of contents, and an
elegant way to do regular paragraphs (<p> elements). The prefix .. is a
little weird, but at least it's explicit. It's possible to be clever
about the whitespace and what it means, but the muforth parser wasn't
really designed to be misused in this way...