[core] More markup generator improvements
Headings now have an "id" attribute and are self-links (by clicking on
one you go to it; makes bookmarking of interior nodes much easier), and
"fenced" <pre> sections now work.

I still want to try to implement non-fenced <pre> sections (any line
starting with whitespace starts a <pre>), but this is trickier.

And unordered lists are still a TODO.