Removed trailing whitespace, everywhere
Not sure why I'm bothering, but...

I started out editing the files, then realized that it would be easier
to use sed.

The device equates files for S08 and AVR had *lots* of trailing ws
because of the way they were created - a simple script that padded (on
the right) the name of each bit in a register. The last bit's padding
ended up being the last characters on the line: trailing spaces.

At least now maybe I'll remember that there is a sed one-liner that will
do this for me. ;-)