[arm] Added some NXP LPC81x support; bootloader might work
I wrote a skeleton bit of code a long time ago to talk to the "LPC
bootloader" (aka ISP -- in system programming -- mode). I originally
wanted to use this with the LPC2138 (I have an Olimex board with this
chip on it) and the LPC2131 (I know someone with a New Micros tini2131
board).

But that work got lost in the seismic shift as everyone ditched the
ARM7TDMI core and went to the Cortex-M series.

NXP's LPC81x are one of these newer Cortex-M cores (Cortex-M0+ in fact)
and the same friend with the tini2131 wanted me to try to get muforth to
talk to the ISP bootloader on this chip.

To make this easier to test, I also added a new "board" type --
target/ARM/board/bare-lpc810.mu4 -- which will load all the right stuff,
including the totally re-written LPC bootloader support code.

I "tested" it against a "peer" version, which I also wrote. Both are
based on the NXP's ISP documentation, which is truly terrible,
misleading, and full of errors. I'll spare you here.

I think it should be possible to connect and read the chip's memory, but
I can't be sure I got everything right.

To try it:

  ./muforth -f target/ARM/board/bare-lpc810.mu4
  isp

and see what happens. "0 du" should show the flash; "@ram du" will show
the RAM.