[msp430] Added beginnings of support for MSP430F5529
I discovered that TI makes a $13 Launchpad based on this chip. One nice
thing about it, other than the low cost, is that the chip has a USB
interface built-in, so it becomes another possible candidate for a
"connect to anything over USB" device.

I still need to get my hands on one, modify the serial chat code, and
then figure out how the USB interface works, but I think this could be a
target worth supporting.