[msp430] Testing multiplication and division (work-in-progress)
The code is a bit of a mess, but I was trying to test my versions of
bit-by-bit multiplication (signed and unsigned) and division (only
unsigned so far). I tested the multiplication operations against the
built-in MPY32 peripheral on the MSP430FR6989, then tried testing
against the host forth's idea of what the right answer should be.

Lastly, I struggled with testing unsigned division, and got something
that kind of works. The biggest struggle: choosing good dividends and
divisors.