[core] Hidden bit is gone - again! Length byte zero means hidden
Another change to the dictionary structure. This gives the same effect
but is simpler, and allows much longer names! Not that the previous
limit of 127 bytes was too short. ;-)

Since I'm using - and really only plan on using - hidden entries as a
way to link vocabulary chains together, using a length of zero rather
than setting the high bit of the length byte will be fine. It could in
theory be used to "erase" words from the dictionary as well, even though
I never do that.