msp430: Added flash support to chat v4 (was fram only)
In adding support for the MSP430F5529 chip (and Launchpad) I realized
that I needed a different way to support flashing the chip -- again
because of the very slow turnaround with the embedded CDC-ACM debug UART
implementation on the Launchpads.

I have checked-in some clock code that is currently ifdef'd out. Instead
I am using the clocks as they are set up by the USB BSL code. That seems
to be an 8MHz clock based on the internal oscillator. There is a crystal
on the Launchpad that I would like to support but that is going to
require some more work.

The way I am currently using this board is as follows:

  * use the USB-BSL to write the chat v4 code into RAM
  * jump to RAM-based chat code
  * flash the Forth kernel code into the start of the flash

This works perfectly, but I would prefer to have the chat code
flash-resident.