[arm] Added basic support for STM32 USB DFU bootloader
In the interest of having *options* other than using the broken ST-LINK
debug interface on STM32 Discovery boards, I decided to add support for
the USB DFU interface available on some boards (eg, on the STM32F3 and
STM32F072 Discovery boards).

I wrote and tested the code using an STM32F3 Discovery board (which is
based on the STM32F303VC).

Only the basic communications have been tested (by reading the flash and
by reading and writing the RAM). I have not tried erasing or programming
the flash, but I have faith that it will work fine. ;-)