[arm] Added STM32 serial bootloader support
Still somewhat of a work-in-progress, but it's possible to connect to a
chip - tested on STM32F051 and STM32F303 - and read flash memory (and
RAM too, but carefully).

One quirk with the STM32 bootloaders - USB DFU and serial both - is that
they consume a good chunk of the RAM when running, and the *user* must
be careful not to try to read or write this RAM area. Using "du" it's
easy to get into trouble.

The good thing is that it's pretty easy to put a RAM- or flash-based
serial chat of our own composition onto the chip and switch to that.
That's a good next step.

Another note: I only tested the memory read. I didn't try to erase or
program any of the flash memory. (And in fact my implementation of the
erase page functionality is incomplete.)

This is a good start, however!