[arm] Completed support for STM32 bootloaders (serial and DFU)
They are now completely wired into the "interactive exploration" feature
of muforth, so it's possible to read and write memory (but not i/o
register contents).

It's now possible to load an image, connect to the target, check the
state of things, erase and flash, and verify that the new contents were
correctly written. All interactively!

Sadly I still have not written a USB or serial chat program. That is
coming soon, hopefully.

An interesting side effect of this work: the Gigadevice GD32VF103 RISC-V
chips use ST's serial and DFU bootloaders! I was able to use this code
to connect to, explore, and erase and program, a GD32VF103CBT6. (I
bought a handful of Sipeed Longan Nano boards from Mouser for exactly
this purpose.)

To be honest, at this point I'm more excited about the Gigadevice parts
than the Freescale/NXP Kinetis or STM32 parts. The ARM Cortex-M is just
too complicated. Even the ARMv6-M parts (like the Cortex-M0 and -M0+)
still have too many details to get right. They are way too much to try
to explain to someone new to microcontrollers - part of muforth's raison
d'etre.

And I still have not finished writing the Thumb32 assembler!!

Happily, though, this work makes possible a cheap and cheerful world
full of RISC-V boards that cost less than $10.