Fixed copyright on several files!
The way I create new .mu4 files is with a Vim autocommand that copies a
"boilerplate" file into the (new) .mu4 buffer. The problem is, it's
using an old copy of muforth that I haven't worked on for a while - with
a "2002-2020" copyright signature!

This was copied into several new files which were committed last month.

This commit brings them up-to-date.