[core, arm] Improved (hopefully) BSD USB support
I'm writing this a bit blind (ie, this is uncompiled and untested code),
but I needed to rewrite a few things anyway and now I've done that.

Minor changes to Linux code; almost all the big changes are in the code
to support FreeBSD version 8 and later, which totally changed how USB
devices are named and enumerated.