[core] du only prints 7-bit ASCII characters
When dumping memory with du, and when printing the contents of memory as
text, only print printable 7-bit ASCII (characters in the range 32 to
126, inclusive).

Unfortunately, UTF-8 screws up the spacing and alignment. I'm not sure
how to solve this, so for now we will have to be old school about it.
:-(