[core] Make -v more useful
-v prints the names of words as they are being defined. It can be useful
to debug crashes that occur while loading code, and also to debug
weird or surprising behavior regarding words not being found.

It used to print a lot of distracting hex (the hex address of the
current chain, and the hex address of the word being defined); now it
prints the stack depth, the *name* of the current vocab chain, and the
name of the word being defined.

Use it as the first command-line flag if you want to see a lot text fly
by!