[arm] Do not load Forth kernel by default!
Now you have to load it yourself where you want it.

I was loading it by default as a "proof of concept" that executing Forth
code from RAM would work, and it made it easy to test target boards.

But once people start developing their own applications, more
flexibility is needed. Also, if you want to, say, connect to a target
board, read the image off it, and save it to a file, reading the kernel
into ram first is pointless and possibly confusing.

It's possible to load the kernel from a file or from the command line.

From the command line:

  ./muforth -f <board> ram -f target/ARM/v6-m/kernel-itc.mu4

to load it into ram region. Replace "ram" with "flash" to load into
flash region.

From a load file:

  ram
  ld target/ARM/v6-m/kernel-itc.mu4

Again, replace ram with flash for flash region.

That's it!