[arm] Big flash programming improvements!
The main points:

* Now possible to program chips - such as the STM32F407 - that use
  non-uniform sector sizes, rather than a uniform page size;

* the RAM buffer used to program the flash is now a variable size, so
  fewer "flash-chunk" operations will occur. Should be faster!

I tested the code on a uniform page device (STM32F303 an an F3-Discovery
board). It worked, but I had to try it *twice*! The first time, the
board hung during the erase. When I tried again, it worked. This mystery
is disturbing. Hopefully all will become clear!