[arm] read-image can read from aliased flash addresses
On STM32 parts, flash starts at 0800_0000, rather that from 0000_0000.

But if you try to read 0000_0000 you will get a copy of the flash
memory, since the under normal circumstances (a normal boot), the chip
hardware maps the flash there.

This change allows read-image to successfully mark the boundaries of the
flash region so that save-image will work properly.