[arm] Raspberry Pi Pico PICOBOOT support!
It's possible to plug in a Raspberry Pi Pico and connect to its
bootloader (PICOBOOT) and read and write the chip's memory!

Unfortunately, it's not possible to read the i/o registers, and it's not
possible to use "du" to read the Flash (because du reads 16 bytes at a
time and the Flash can only be read 256 bytes at a time).

Some fun things to do:

Plug in the board. Then do

  pico
  0 du

You'll see the contents of the boot ROM.

  @ram du

will show the contents of the RAM.

  @flash #flash read-image
  save-image pico-factory.img

will read the entire contents of the flash and write it to a file in
muforth/mu that you can either hexdump, disassemble using the tools you
like, or even load it back into muforth and look at it there!

Enjoy! There is more on the way.