[arm] Booting example code for STM32F072 Discovery board
This is much like the code I recently committed for the F303 Discovery
board. It contains:

- a vector table
- a Forth kernel
- some simple init code

The init code sets up the clock to run from the on-chip HSI48 48 MHz
clock, sets up Port C to be clocked, sets the MODER bits for Port C, and
flashes the LEDs once.

Use it as another example of how to create a flashable image for these
boards.