[arm] Enjoy executing Forth (from RAM) on your RasPi Pico!
It's now possible to copy a small RAM-resident Forth kernel to the RAM
on the Raspberry Pi Pico board, and execute code interactively!

For now this is the only way to read the i/o registers. Oddly, it's not
possible to read the flash this way (ie, using @). It just returns 0.

Before being able to do much exploration, we need an equates file for
the board!

But I'm pleased that it was relatively easy to get this to work.