[arm] Make @ and ! (and c/h variants) work better during interaction
In the past it was very easy to segfault, either before connecting to a
target, or when connected to a target that doesn't have memory words
defined in its own Forth kernel.

The reason? @ and ! would get executed on a *target* address, and
because this is outside of muforth's heap, the memory operation would
segfault.

The solution? I've defined memory operators that live in the .meta.
chain - so they are found while building a target image and while
chatting with a live target - that are smart about target addresses.

@ h@ and c@, when *not* chatting, simply return the contents of the host
image for that target address.

@ h@ and c@, when chatting, read the memory of the target, using the
connected debug interface.

! h! and c!, when *not* chatting, simply write to the image for that
target address.

! h! and c!, when chatting, write to *both* the image and target memory.

One other addition in this commit: a third region, called "bogus", that
gets mapped for any address outside of flash or ram.

If you try to read an i/o register when not chatting, you will see the
value "0bee_f00d instead. But you will not segfault!!